Update 20160405
Exercise 1.16.  Design a procedure that evolves an iterative exponentiation process that uses successive squaring and uses a logarithmic number of steps, as does fast-expt. (Hint: Using the observation that (bn/2)2 = (b2)n/2, keep, along with the exponent n and the base b, an additional state variable a, and define the state transformation in such a way that the product a bn is unchanged from state to state. At the beginning of the process a is taken to be 1, and the answer is given by the value of a at the end of the process. In general, the technique of defining an invariant quantity that remains unchanged from state to state is a powerful way to think about the design of iterative algorithms.)


one point worth paying attention to , the iteration  ends at state n = 0,  but n = 1 is the path which can not be bypassed, whatever n is even or odd,

the step is key: f( (* a b)  b (- n 1)), as one could easily figure out, you get n = 0, you have to get n = 1 at first, unless you get n = 0 t